Add files via upload
a.

The key elements of the application include user, items and cart. The user page would allow the user to login thr page. The item page would display the items. The cart page would allow the users to add or remove items from the cart.

A state-managed solution would be appropriate for this application, as it would allow the user to securely log in to their account and view their data without having to re-enter their credentials every time. A Model-View-Controller (MVC) architecture would be the best choice for this application, as it would provide a structured framework for the application to be built upon

b. The environments used in this are HTML5, JDK8, Apache Tomcat8, JDBC API. We cannot avoid writing code as it runs on MVC pattern.
c. Step 1: On “add to cart” action the sessionStorage object builds the purchased item array.
Step 2: Then, it gets the buyer’s payment details on a checkout form.
Step 3: Can renders payment options like PayPal with the request parameter array. This array contains purchased items and the buyer’s payment details.
Step 4: An alternate to step 3, that is, an email checkout. This suits the shopping cart not dealing with instant payment on the flow.